+++ tests/test_network/vnic/test_network_vnic.py
@@ -1,9 +1,10 @@
 from unittest import TestCase
-from pyVmomi import vim
 from mock import Mock, MagicMock
 
-from vCenterShell.network import *
-from vCenterShell.network.vnic.vnic_common import *
+
+from pyVmomi import vim
+from common.vcenter.vmomi_service import *
+from vCenterShell.network.vnic.vnic_service import VNicService
 
 
 class TestNetwork(TestCase):
@@ -14,32 +15,33 @@ class TestNetwork(TestCase):
         vm = Mock()
         vm.ReconfigVM_Task = lambda x: isinstance(x,  vim.vm.ConfigSpec)
 
-        res = vm_reconfig_task(vm, [])
+        api_wrapper = pyVmomiService(Mock, Mock())
+        res = api_wrapper.vm_reconfig_task(vm, [])
         self.assertTrue(res)
 
     def test_compose_empty(self):
-        nicspec = vnic_compose_empty()
+        nicspec = VNicService.vnic_compose_empty()
         self.assertTrue(isinstance(nicspec, vim.vm.device.VirtualDeviceSpec))
         self.assertTrue(isinstance(nicspec.device, vim.vm.device.VirtualVmxnet3))
         self.assertTrue(isinstance(nicspec.device.connectable, vim.vm.device.VirtualDevice.ConnectInfo))
 
     def test_device_attahed_to_network_standard(self):
 
-        self.assertFalse(device_is_attached_to_network(None, None))
+        self.assertFalse(VNicService.device_is_attached_to_network(None, None))
 
         network_name = "TEST"
         device = Mock()
         device.backing = Mock()
         device.backing.network = Mock()
         device.backing.network.name = network_name
-        self.assertTrue(device_is_attached_to_network(device, network_name))
+        self.assertTrue(VNicService.device_is_attached_to_network(device, network_name))
 
         network = Mock(spec=vim.Network)
         network.name = "xnet"
         nicspec = Mock()
 
         nicspec.device = device
-        res = vnic_attach_to_network_standard(nicspec, network)
+        res = VNicService.vnic_attach_to_network_standard(nicspec, network)
         self.assertEquals(res.device.backing.network.name, "xnet")
 
     def test_device_attahed_to_network_distributed(self):
@@ -50,7 +52,7 @@ class TestNetwork(TestCase):
         hasattr(device.backing, "network")
         del device.backing.network
         device.backing.port.portgroupKey = network_name
-        self.assertTrue(device_is_attached_to_network(device, network_name))
+        self.assertTrue(VNicService.device_is_attached_to_network(device, network_name))
 
         port_group = Mock(spec=vim.dvs.DistributedVirtualPortgroup)
         port_group.key = "group_net"
@@ -58,7 +60,7 @@ class TestNetwork(TestCase):
         nicspec = Mock()
 
         nicspec.device = device
-        res = vnic_attach_to_network_distributed(nicspec, port_group)
+        res = VNicService.vnic_attach_to_network_distributed(nicspec, port_group)
         self.assertEquals(res.device.backing.port.portgroupKey, "group_net")
 
 
@@ -66,7 +68,7 @@ class TestNetwork(TestCase):
         vm = Mock()
         vm.ReconfigVM_Task = lambda x: isinstance(x,  vim.vm.ConfigSpec)
         nicspec = Mock()
-        res = vnic_add_to_vm_task(nicspec, vm)
+        res = VNicService.vnic_add_to_vm_task(nicspec, vm)
         self.assertIsNone(res)
 
         # nicspec = Mock(spec=vim.vm.device.VirtualDeviceSpec)
@@ -78,12 +80,12 @@ class TestNetwork(TestCase):
         nicspec = Mock()
         nicspec.device = Mock()
         connect_status = True
-        nicspec = vnic_set_connectivity_status(nicspec, connect_status)
+        nicspec = VNicService.vnic_set_connectivity_status(nicspec, connect_status)
         self.assertEquals(nicspec.device.connectable.connected, connect_status)
 
 
     def test_vnic_is_attachet_to_network(self):
         nicspec = Mock()
         nicspec.device = Mock()
-        res = vnic_is_attachet_to_network(nicspec, Mock())
+        res = VNicService.vnic_is_attachet_to_network(nicspec, Mock())
         self.assertFalse(res)
